requirements.txt
pycrypto
scapy